+++ src/hooks/customlists/useCustomList.js
@@ -5,6 +5,7 @@ export default function useCustomList(listName) {
     const { gameList, isAchievementUnlocker, isCardFarming } = useContext(AppContext);
     const [list, setList] = useState([]);
     const [searchTerm, setSearchTerm] = useState('');
+    const [showInList, setShowInList] = useState(false);
     const [visibleGames, setVisibleGames] = useState(50);
     const containerRef = useRef(null);
 
@@ -72,6 +73,8 @@ export default function useCustomList(listName) {
         containerRef,
         searchTerm,
         setSearchTerm,
+        showInList,
+        setShowInList,
         handleAddGame,
         handleRemoveGame,
         updateListOrder