+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [![Downloads](https://img.shields.io/github/downloads/zevnda/steam-game-idler/1.7.4/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/zevnda/steam-game-idler/releases/download/1.7.4/Steam.Game.Idler_1.7.4_x64_en-US.msi)
 
 ## Changelog
-- Miscellaneous UI styling changes
+- Added an `in list` button to the `edit list modal` to filter by games already in the list
+- Made the search bar in `edit list modal` clearable